requirements.txt
# SoC Validation Test Infrastructure - Python Dependencies

# Core Dependencies
prefect==3.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database and Storage
redis==5.0.1
sqlalchemy==2.0.25
alembic==1.13.1

# Async and Networking
httpx==0.26.0
aiofiles==23.2.1
telnetlib3==2.0.4
paramiko==3.4.0  # For SSH/SFTP support

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
python-multipart==0.0.6

# Monitoring and Logging
structlog==24.1.0
prometheus-client>=0.20.0

# Notifications
slack-sdk==3.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0

# Code Quality
black==23.12.1
ruff==0.1.11
mypy==1.8.0
pre-commit==3.6.0